Remove old files, move.py to master dir
