+++ output_analysis/output_title_stop_words_20220930.csv
@@ -1,0 +1,196 @@
+,0
+0,these
+1,out
+2,yours
+3,being
+4,between
+5,again
+6,both
+7,doesn
+8,yourselves
+9,at
+10,with
+11,m
+12,itself
+13,didn't
+14,his
+15,there
+16,those
+17,some
+18,as
+19,ve
+20,you've
+21,haven
+22,mustn't
+23,won't
+24,needn't
+25,"
+"
+26,them
+27,you
+28,you'll
+29,up
+30,be
+31,my
+32,hasn't
+33,y
+34,if
+35,for
+36,over
+37,other
+38,to
+39,your
+40,its
+41,that'll
+42,each
+43,yourself
+44,results:
+45,an
+46,hers
+47,t
+48,don
+49,does
+50,while
+51,background:
+52,ourselves
+53,re
+54,introduction:
+55,most
+56,should've
+57,didn
+58,do
+59,don't
+60,he
+61,href
+62,objective:
+63,hadn't
+64,our
+65,i
+66,same
+67,below
+68,is
+69,are
+70,were
+71,which
+72,aren't
+73,haven't
+74,but
+75,above
+76,few
+77,isn't
+78,wouldn
+79,very
+80,her
+81,their
+82,have
+83,how
+84,did
+85,or
+86,against
+87,had
+88,doesn't
+89,wasn't
+90,ours
+91,having
+92,it
+93,themselves
+94,you're
+95,couldn
+96,just
+97,by
+98,can
+99,who
+100,during
+101,more
+102,before
+103,she
+104,has
+105,it's
+106,she's
+107,nor
+108,him
+109,o
+110,aren
+111,about
+112,d
+113,weren't
+114,once
+115,on
+116,shan
+117,we
+118,this
+119,such
+120,himself
+121,not
+122,from
+123,no
+124,mustn
+125,ain
+126,own
+127,than
+128,theirs
+129,you'd
+130,a
+131,into
+132,that
+133,why
+134,wasn
+135,conclusions:
+136,where
+137,then
+138,doing
+139,what
+140,was
+141,hadn
+142,so
+143,after
+144,here
+145,needn
+146,should
+147,shouldn
+148,under
+149,they
+150,won
+151,through
+152,of
+153,mightn't
+154,isn
+155,whom
+156,myself
+157,weren
+158,me
+159,been
+160,down
+161,methods:
+162,any
+163,too
+164,in
+165,purpose:
+166,further
+167,the
+168,off
+169,only
+170,when
+171,couldn't
+172,shan't
+173,ll
+174,hasn
+175,s
+176,herself
+177,am
+178,ma
+179,br
+180,wouldn't
+181,all
+182,will
+183,because
+184,mightn
+185,and
+186,now
+187,until
+188,shouldn't
+189,care
+190,aged
+191,health
+192,government
+193,department
