--- text_mine/final_stop_words_abstract.csv
@@ -1,189 +1,0 @@
-,0
-0,over
-1,are
-2,to
-3,my
-4,by
-5,objective:
-6,doing
-7,own
-8,was
-9,i
-10,on
-11,so
-12,who
-13,not
-14,can
-15,conclusions:
-16,y
-17,hadn
-18,above
-19,hers
-20,he
-21,is
-22,out
-23,couldn't
-24,it
-25,yourselves
-26,now
-27,aren
-28,doesn
-29,href
-30,what
-31,isn
-32,when
-33,wouldn
-34,did
-35,herself
-36,you'll
-37,and
-38,such
-39,weren
-40,hasn
-41,have
-42,has
-43,during
-44,re
-45,again
-46,yourself
-47,ourselves
-48,myself
-49,this
-50,against
-51,until
-52,after
-53,they
-54,there
-55,mightn
-56,an
-57,their
-58,haven't
-59,should
-60,why
-61,up
-62,both
-63,from
-64,your
-65,other
-66,whom
-67,through
-68,same
-69,hadn't
-70,br
-71,between
-72,for
-73,him
-74,because
-75,t
-76,here
-77,you'd
-78,some
-79,weren't
-80,into
-81,nor
-82,haven
-83,introduction:
-84,wouldn't
-85,had
-86,very
-87,any
-88,at
-89,you've
-90,under
-91,should've
-92,too
-93,yours
-94,being
-95,that
-96,where
-97,while
-98,won
-99,all
-100,theirs
-101,her
-102,d
-103,ours
-104,its
-105,background:
-106,ma
-107,mustn
-108,than
-109,ain
-110,most
-111,results:
-112,themselves
-113,didn't
-114,our
-115,hasn't
-116,mightn't
-117,then
-118,be
-119,won't
-120,having
-121,further
-122,no
-123,those
-124,you
-125,only
-126,it's
-127,will
-128,methods:
-129,them
-130,as
-131,just
-132,of
-133,me
-134,ll
-135,aren't
-136,needn
-137,shan't
-138,shouldn
-139,his
-140,himself
-141,ve
-142,didn
-143,you're
-144,in
-145,mustn't
-146,wasn
-147,been
-148,the
-149,each
-150,down
-151,o
-152,do
-153,am
-154,does
-155,shan
-156,needn't
-157,s
-158,these
-159,few
-160,but
-161,once
-162,or
-163,below
-164,more
-165,wasn't
-166,were
-167,she's
-168,which
-169,purpose:
-170,if
-171,itself
-172,she
-173,don't
-174,isn't
-175,how
-176,doesn't
-177,with
-178,that'll
-179,before
-180,don
-181,shouldn't
-182,off
-183,m
-184,about
-185,a
-186,couldn
-187,we
